+++ resources/components/laraknife/icons/task.blade.php
@@ -1,0 +1,2 @@
+@props(['class' => ''])
+<i class="bi bi-list-task {{$class}}"></i>