+++ resources/helpers/EmailHelper.php
@@ -3,6 +3,7 @@ namespace App\Helpers;
 
 use App\Helpers\ViewHelper;
 use App\Mail\ForgottenPassword;
+use App\Mail\NoteNotification;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Mail\Mailables\Address;
 
@@ -13,7 +14,12 @@ class EmailHelper
         switch ($name) {
             case 'user.forgotten':
                 Mail::to($to)->send(
-                    new ForgottenPassword($snippets['link'])
+                    new ForgottenPassword($snippets)
+                );
+                break;
+            case 'note.notification':
+                Mail::to($to)->send(
+                    new NoteNotification($snippets)
                 );
                 break;
             default: