+++ resources/views/mails/note-notification.blade.php
@@ -1,0 +1,7 @@
+<p>{{ __('Hello')}} {{ $name }}, </p>
+<p>{{__('The responsibility has been transferred to you')}} ({{ $from }}).<br/>
+<a href="{{$link}}">{{$link}}</a></p>
+<hr>
+<h3>{{$title}}</h3>
+<p>{{$contents}}
+</p>