+++ templates/database/seeders/NoteSeeder.php
@@ -2,6 +2,7 @@
 
 namespace Database\Seeders;
 
+use App\Models\SProperty;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Database\Console\Seeds\WithoutModelEvents;
@@ -13,55 +14,20 @@ class NoteSeeder extends Seeder
      */
     public function run(): void
     {
-        DB::delete('DELETE FROM sproperties WHERE id>=1011 and id < 1100');
-        DB::table('sproperties')->insert([
-            'id' => 1011,
-            'scope' => 'notestatus',
-            'name' => 'open',
-            'order' => '10',
-            'shortname' => 'O'
-        ]);
-        DB::table('sproperties')->insert([
-            'id' => 1012,
-            'scope' => 'notestatus',
-            'name' => 'closed',
-            'order' => '20',
-            'shortname' => 'C'
-        ]);
-        DB::table('sproperties')->insert([
-            'id' => 1051,
-            'scope' => 'category',
-            'name' => 'standard',
-            'order' => '10',
-            'shortname' => 'Std'
-        ]);
-        DB::table('sproperties')->insert([
-            'id' => 1052,
-            'scope' => 'category',
-            'name' => 'private',
-            'order' => '20',
-            'shortname' => 'P'
-        ]);
-        DB::table('sproperties')->insert([
-            'id' => 1053,
-            'scope' => 'category',
-            'name' => 'work',
-            'order' => '30',
-            'shortname' => 'W'
-        ]);
-        DB::table('sproperties')->insert([
-            'id' => 1091,
-            'scope' => 'visibility',
-            'name' => 'public',
-            'order' => '10',
-            'shortname' => 'PU'
-        ]);
-        DB::table('sproperties')->insert([
-            'id' => 1092,
-            'scope' => 'visibility',
-            'name' => 'private',
-            'order' => '20',
-            'shortname' => 'PV'
-        ]);
+        // int $id, string $scope, string $name, int $order, string $shortname
+        SProperty::insertIfNotExists(1011, 'notestatus', 'open', '10', 'O');
+        SProperty::insertIfNotExists( 1012, 'notestatus', 'closed', '20', 'C' );
+
+        SProperty::insertIfNotExists( 1051, 'category', 'standard', '10', 'Std' );
+        SProperty::insertIfNotExists( 1052, 'category', 'private', '20', 'P' );
+        SProperty::insertIfNotExists( 1053, 'category', 'work', '30', 'W' );
+        SProperty::insertIfNotExists( 1054, 'category', 'task', '40', 'T' );
+        
+        SProperty::insertIfNotExists( 1091, 'visibility', 'public', '10', 'PU' );
+        SProperty::insertIfNotExists( 1092, 'visibility', 'private', '20', 'PV' );
+
+        SProperty::insertIfNotExists(1301, 'task', 'description', '10', 'D');
+        SProperty::insertIfNotExists(1302, 'task', 'snaketext', '20', 'S');
+        SProperty::insertIfNotExists(1303, 'task', 'cloze', '30', 'C');
     }
 }